% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.R
\name{create_data}
\alias{create_data}
\title{Generate simulated observational data}
\usage{
create_data(DGP, n = 1)
}
\arguments{
\item{DGP}{a data generating process, which is the list output of the dgp() function}

\item{n}{the number of desired samples}
}
\description{
Generates N tuples from the joint distribution of P(X,W,Y).
Returns a list with two elements: a dataframe called data and a dataframe 
called aux_data containing true expecations conditional on X.
}
