% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.R
\name{dgp}
\alias{dgp}
\title{Generate simulated observational data}
\usage{
dgp()
}
\arguments{
\item{X}{a list of distribution objects that generate a covariate vector}

\item{f_W_x}{a function of x that returns the conditional distribution W|X=x}

\item{f_Y_xw}{a function of x, w that returns the conditional distribution Y|X=x,W=w}

\item{f_C_xw}{a function of x, w that returns the conditional distribution C|X=x,W=w}
}
\description{
Generates N tuples from the joint distribution of P(X,W,Y).
Returns a list with two elements: a dataframe called data and a dataframe 
called aux_data containing true expecations conditional on X.
}
