% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_hte_cv.R
\name{get_errors}
\alias{get_errors}
\title{Estimates estimated cv errors and true test errors (the latter via true values in aux_data).}
\usage{
get_errors(cv_estimates, test_estimates, aux_data)
}
\arguments{
\item{cv_estimates}{cv_estimates}

\item{test_estimates}{test_estimates}

\item{aux_data}{aux_data}
}
\description{
Estimates estimated cv errors and true test errors (the latter via true values in aux_data).
}
