% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_hte_cv.R
\name{get_estimates}
\alias{get_estimates}
\title{Estimates conditonal mean regression models on the data via caret. 
Does cross-estimation on the training data and runs each model trained on the full training set on the test set.
Returns all out-of-sample estimates from each model}
\usage{
get_estimates(data, models, cv_index, test_index)
}
\arguments{
\item{data}{data}

\item{models}{models}

\item{cv_index}{cv_index}

\item{test_index}{test_index}
}
\description{
Estimates conditonal mean regression models on the data via caret. 
Does cross-estimation on the training data and runs each model trained on the full training set on the test set.
Returns all out-of-sample estimates from each model
}
