% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_hte_cv.R
\name{get_metrics}
\alias{get_metrics}
\title{Estimates estimated cv metrics and true test metrics (the latter via true values in aux_data).}
\usage{
get_metrics(cv_estimates, test_estimates, aux_data)
}
\arguments{
\item{cv_estimates}{cv_estimates}

\item{test_estimates}{test_estimates}

\item{aux_data}{aux_data}
}
\description{
Estimates estimated cv metrics and true test metrics (the latter via true values in aux_data).
}
