% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.R
\name{powers_DGPs}
\alias{powers_DGPs}
\title{Simulations from Powers et al. 2018}
\usage{
powers_DGPs()
}
\description{
Creates a set of data generating processes from Powers et al 2018 that can be passed to the
create_data function to simulate observational data
}
