% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_hte_cv.R
\name{setup_data}
\alias{setup_data}
\title{Prepares simulated data for experiments}
\usage{
setup_data(DGP, n, training_percent, n_folds)
}
\arguments{
\item{DGP}{a list created by a call to dgp()}

\item{n}{the number of samples desired}

\item{training_percent}{how much data to use for training}

\item{n_folds}{the number of folds to use for treatment effect cross-validation}
}
\description{
Prepares simulated data for experiments
}
