% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_scenarios.R
\name{dgp}
\alias{dgp}
\title{Generate simulated observational data}
\usage{
dgp(covariate_fun, propensity_fun, mean_fun, effect_fun, sigma)
}
\arguments{
\item{covariate_fun}{a function that randomly takes n samples from P(X)}

\item{propensity_fun}{a vectorized function of a single vector-valued arugment x that defines E[W|X=x]}

\item{mean_fun}{a vectorized function of a vector-valued arugment x that defines E[Y|X=x]}

\item{effect_fun}{a vectorized function of a single vector-valued arugment x that defines E[Y|W=1,X=x] - E[Y|W=0,X=x]}

\item{sigma}{standard deviation of the normally distributed zero-mean additive noise to Y}
}
\description{
Generates N tuples from the joint distribution of P(X,W,Y).
Returns a list with two elements: a dataframe called data and a dataframe 
called aux_data containing true expecations conditional on X.
}
