% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_hte_cv.R
\name{setup_data}
\alias{setup_data}
\title{Prepares simulated data for experiments}
\usage{
setup_data(DGP, n_train, n_test, n_folds)
}
\arguments{
\item{DGP}{a list created by a call to dgp()}

\item{n_train}{number of samples to train and validate on}

\item{n_test}{number of samples to test on}

\item{n_folds}{the number of folds to use for treatment effect cross-validation}
}
\description{
Prepares simulated data for experiments
}
